PLX-1008 moved lambda sources into internal src
