+++ operations/build-release/bundle-zip.sh
@@ -27,13 +27,13 @@ zip $RELEASE_BASE_ZIP_FILENAME ./package-lock.json
 source ./operations/utils/aws-release.utils.sh --delete-current-local-release
 
 CURRENT_DIRECTORY=$(pwd)
-AWS_LAMBDA_SRC_APP_DIRECTORY="$CURRENT_DIRECTORY/src/app/awsLambdas"
-AWS_LAMBDA_DIST_DIRECTORY="$AWS_LAMBDA_SRC_APP_DIRECTORY/dist"
-AWS_LAMBDA_DIST_APP_DIRECTORY="$AWS_LAMBDA_SRC_APP_DIRECTORY/dist/app/awsLambdas"
+AWS_LAMBDA_APP_DIRECTORY="$CURRENT_DIRECTORY/src/app/awsLambdas"
+AWS_LAMBDA_DIST_DIRECTORY="$AWS_LAMBDA_APP_DIRECTORY/dist"
+AWS_LAMBDA_DIST_APP_DIRECTORY="$AWS_LAMBDA_APP_DIRECTORY/dist/app/awsLambdas"
 
 # Add files to 'dist' directory
-cp "$AWS_LAMBDA_SRC_APP_DIRECTORY/package.json" $AWS_LAMBDA_DIST_APP_DIRECTORY
-cp "$AWS_LAMBDA_SRC_APP_DIRECTORY/package-lock.json" $AWS_LAMBDA_DIST_APP_DIRECTORY
+cp "$AWS_LAMBDA_APP_DIRECTORY/package.json" $AWS_LAMBDA_DIST_APP_DIRECTORY
+cp "$AWS_LAMBDA_APP_DIRECTORY/package-lock.json" $AWS_LAMBDA_DIST_APP_DIRECTORY
 
 printf "\n\nCreating lbapi1 release zip file '"$RELEASE_LBAPI1_ZIP_FILENAME"'...\n"
 rm -f $RELEASE_LBAPI1_ZIP_FILENAME 2>/dev/null
