+++ src/app/awsLambdas/src/api/__tests__/__data__/lambdaEventFromApiGw.data.json
