+++ src/app/awsLambdas/src/api/__tests__/handler.test.ts
