+++ src/app/awsLambdas/src/api/types.ts
